Week 4 - Final
Adjustments to the javascript to remove the dots and move the
previous/next buttons toward the bottom. Fixed the issue with not
automatically viewing the slideshow right away (before hitting the dots
the viewer saw all the articles at once instead of in sequence). Added
a headshot and quote of the author to give a better intro of the
site.Added the corrections text file.